move shared MustacheTag/RawMustacheTag logic
